Conflicts:
	src/crispy.h
	src/i_video.c
	textscreen/txt_sdl.c

+++ src/doom/p_maputl.c
@@ -470,7 +470,7 @@ P_BlockLinesIterator
   boolean(*func)(line_t*) )
 {
     int			offset;
-    long*		list; // [crispy] BLOCKMAP limit
+    int64_t*		list; // [crispy] BLOCKMAP limit
     line_t*		ld;
 	
     if (x<0
@@ -864,7 +864,7 @@ static void InterceptsOverrun(int num_intercepts, intercept_t *intercept)
 
     InterceptsMemoryOverrun(location, intercept->frac);
     InterceptsMemoryOverrun(location + 4, intercept->isaline);
-    InterceptsMemoryOverrun(location + 8, (int) intercept->d.thing);
+    InterceptsMemoryOverrun(location + 8, (intptr_t) intercept->d.thing);
 }
